Fixed a bug

+++ LogicGatesInCPP.cpp
@@ -6,7 +6,7 @@ int main() {
 	Numeric<> six(6);
 	Numeric<> five(5);
 
-	std::cout << Or<>(&Numeric<>(6), &Numeric<>(5)).operation() << std::endl;
+	std::cout << static_cast<int>(Or<>(&Numeric<>(6), &Numeric<>(5)).operation()) << std::endl;
 
 	return 0;
 }
